--- libraries/EPOS2Control/EPOS2Control.h
@@ -1,7 +1,0 @@
-#ifndef EPOS2CONTROL_H
-#define EPOS2CONTROL_H
-
-#include <mcp_can.h>
-#include <Arduino.h>
-
-#endif
