--- libraries/EPOS2ControlMsgUtils/EPOS2ControlMsgUtils.h
@@ -1,6 +1,0 @@
-#ifndef EPOS2CONTROLMESSAGEUTILS_H
-#define EPOS2CONTROLMESSAGEUTILS_H
-#include "Arduino.h"
-
-
-#endif  // EPOS2CONTROLMESSAGEUTILS_H
