--- libraries/EPOS2ControlSetup/EPOS2ControlSetup.h
@@ -1,14 +1,0 @@
-#ifndef EPOS2CONTROLSETUP_H
-#define EPOS2CONTROLSETUP_H
-
-#include <mcp_can.h>
-#include <Arduino.h>
-
-// void setupPDOs();
-// void setupRPDO1();
-// void setupTPDO1();
-// void setupHearbeat();
-// void setupVelocityMode();
-// void setInitialVals();
-
-#endif // EPOS2CONTROLSETUP_H
