+++ libraries/MatlabSerial/MatlabSerial.h
@@ -1,6 +1,7 @@
 #ifndef MATLABSERIAL_H
 #define MATLABSERIAL_H
 
+#include "MatlabSerial.h"
 #include "Arduino4Tetis.h"
 
 class KinematicSystem;
