--- libraries/TetisKinematics/TetisKinematics.h
@@ -1,9 +1,0 @@
-#ifndef TETISKINEMATICS_H
-#define TETISKINEMATICS_H
-
-#include <mcp_can.h>
-#include <Arduino.h>
-
-
-
-#endif //TETISKINEMATICS_H
